test(sanitize): split examples and benchmarks
